tasks implemented: crud, h2console, exception, swagger doc, seed data

+++ src/main/java/com/stackroute/service/TrackServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.service;
 
 import com.stackroute.domain.Track;
+import com.stackroute.exceptions.TrackAlreadyExistsException;
+import com.stackroute.exceptions.TrackNotFoundException;
 import com.stackroute.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,10 +19,14 @@ public class TrackServiceImpl implements TrackService {
         this.trackRepository=trackRepository;
     }
     @Override
-    public Track saveTrack(Track track) {
-
-        Track savedTrack = trackRepository.save(track);
-        return savedTrack;
+    public Track saveTrack(Track track) throws TrackAlreadyExistsException {
+        if (trackRepository.existsById(track.getId())){
+            throw new TrackAlreadyExistsException("Track Already exists");
+        }
+        else {
+            Track savedTrack = trackRepository.save(track);
+            return savedTrack;
+        }
     }
 
     @Override
@@ -35,8 +41,13 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public void deleteTrack(Track track) {
-        trackRepository.deleteById(track.getId());
+    public void deleteTrack(Track track) throws TrackNotFoundException {
+        if (trackRepository.existsById(track.getId())) {
+            trackRepository.deleteById(track.getId());
+        }
+        else {
+            throw new TrackNotFoundException("Track doesn't exist!");
+        }
     }
 
     @Override
